Re-target project to .NET Standard 2.1
Unfortunately this project has not been updated for some time so it has
not kept pace with the change to .NET Standard and now to .NET Core.

In this commit we have used a multi-targeting approach so that it can be
consumed from either .NET Core or by .NET Framework 4.7.2 or above.

As NuGet packaging can be done using attributes in the library and
different targets are easily packed with the distributable NuGet package
we can remove much of the NuGet specific code in the repository also.