Replace deprecated reflection API
The 'LambdaExpression.CompileToMethod' method has been deprecated in
.NET Standard 2.1 and in .NET Core. This omission means that we need to
find an alternative way to generate the IL for methods that convert
values when serializing parameters.

The approach I have taken here is something of a hack. Essentially we
create a field where we can store the lambda expression after it has
been compiled to a delegate. Next we generate a method that loads the
field and invokes the delegate. The field value is set once the type has
been created and now the method will run the correct lambda expression.

Generating converters for interfaces that have generic methods is
considered outside the scope of this change.